+++ src/MobileMenuModule.php
@@ -53,55 +53,37 @@ class MobileMenuModule extends \Module
      */
     protected function compile()
     {
-        $position = $this->mobile_menu_position;
+        $this->Template->trigger        = $this->mobile_menu_trigger;
+        $this->Template->html           = $this->mobile_menu_html;
+        $this->Template->position       = $this->mobile_menu_position;
+        $this->Template->size           = (int) $this->mobile_menu_size;
+        $this->Template->overlay        = $this->mobile_menu_overlay;
+        $this->Template->offCanvas      = $this->mobile_menu_offCanvas;
+        $this->Template->animation      = $this->mobile_menu_animation;
+        $this->Template->animationSpeed = (int) $this->mobile_menu_animationSpeed;
+        $this->Template->mediaQuery     = $this->mobile_menu_mediaQuery;
 
-        // Backwards compatibility
-        if ($position == '') {
-            $position = 'left';
-        }
-
-        $this->Template->trigger = $this->mobile_menu_trigger;
-        $this->Template->html = $this->mobile_menu_html;
-        $this->Template->position = $position;
-        $this->Template->size = $this->mobile_menu_size;
-
-        $triggerClass = '';
+        $breakPoint = 0;
 
         // Display on phones
         if ($this->mobile_menu_phones) {
-            $triggerClass .= ' display_phones';
+            $breakPoint = 767;
         }
 
         // Display on tablets
         if ($this->mobile_menu_tablets) {
-            $triggerClass .= ' display_tablets';
+            $breakPoint = 991;
         }
 
-        $this->Template->triggerClass = $triggerClass;
-        $this->Template->menuStyle = array();
-
-        // Add the mobile menu custom size
-        if ($this->mobile_menu_size) {
-            switch ($position) {
-                case 'left':
-                case 'right':
-                    $property = 'width';
-                    break;
-
-                case 'top':
-                case 'bottom':
-                    $property = 'height';
-                    break;
-            }
-
-            $this->Template->menuStyle = array(
-                'property' => $property,
-                'value'    => $this->mobile_menu_size . '%',
-            );
+        // Set a custom break point
+        if ($this->mobile_menu_breakpoint) {
+            $breakPoint = (int) $this->mobile_menu_breakpoint;
         }
+
+        $this->Template->breakPoint = $breakPoint;
 
         // Add assets
         $GLOBALS['TL_CSS']['mobileMenu'] = 'system/modules/mobile_menu/assets/css/mobile-menu.min.css||static';
-        $GLOBALS['TL_JAVASCRIPT']['mobileMenu'] = 'system/modules/mobile_menu/assets/js/mobile-menu.min.js|static';
+        $GLOBALS['TL_JAVASCRIPT']['mobileMenu'] = 'system/modules/mobile_menu/assets/js/mobile-menu.jquery.min.js|static';
     }
 }